Split common.go into smaller files.
